multi: make lib/types the canonical TxProof codec location
Move the TxProof TLV codec into lib/types as the single source of
truth, eliminating the duplicate in db/proof_codec.go. The db package
now imports types.SerializeTxProof and types.DeserializeTxProof
instead of maintaining its own copy. This removes the maintenance
risk of the two implementations diverging.
